First release - functional but a mess
comitting the functional release. It's a bit of a mess, need to clean up things into functions and whatnot.

monster_parse.py was written to parse Fight Club 5e xml files into a discord message for a discord bot.

scratch.py was written to build off of that and reformat the xml files into the same (or similar) json format as used in bobifles scripts. I also changed the tokens.py code because I couldn't get the lib or POI tokens to build, but I successfully got all the tokens from my custom XML to build (which I can't include for legal reasons).

Also, this just makes all the tokens into a bear because I don't have bobifles imglib.